% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.plot_one_interaction}
\alias{.plot_one_interaction}
\title{Network visualization}
\usage{
.plot_one_interaction(adjacency_matrix, n, color, name)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{...}{Optional parameters.}
}
\description{
Creates a graph based on interactions provided
}
