% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fc}
\alias{fc}
\title{Example fcoex object}
\format{
An object of class \code{fcoex}
}
\usage{
fc
}
\description{
Example fcoex object processed from the mini_pbmc3k dataset.
}
\examples{
data(fc)
fc
}
\keyword{datasets}
