% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{class}
\name{fcoex-class}
\alias{fcoex-class}
\title{An S4 class to represent the fcoex analysis.}
\description{
An S4 class to represent the fcoex analysis.
}
\section{Slots}{

\describe{
\item{\code{expression}}{Normalized gene expression table from single-cells \code{data.frame}.}

\item{\code{discretized_expression}}{Discretized gene expression table from single-cells \code{data.frame}.}

\item{\code{target}}{Original target classes for the cells (\code{factor}).}

\item{\code{selected_genes}}{Character \code{vector} containing the names of genes  selected for analysis}

\item{\code{module_list}}{\code{list} containing genes in each module.}

\item{\code{adjacency}}{\code{data.frame} containing the adjacency table for the selected genes.}

\item{\code{interaction_plot}}{list of ggplot graphs with module gene interactions.}

\item{\code{new_clusters}}{\code{list} containing gene interactions present in modules.}

\item{\code{mod_colors}}{character \code{vector} containing colors associated with each network module.}

\item{\code{ora}}{Over-representation analysis results \code{data.frame}.}

\item{\code{barplot_ora}}{list of ggplot graphs with over-representation analysis results per module.}

\item{\code{mod_idents}}{Identities of cells based on each co-expression module. Determined by the "recluster" method}
}}

\examples{
# Get example expression data
data(expr0)
# Initialize fcoex object with expression
fc <- new("fcoex", expression=expr0)
}
