% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R, R/find_cbf_modules.R
\name{find_cbf_modules}
\alias{find_cbf_modules}
\title{find_cbf_modules}
\usage{
find_cbf_modules(discretized_exprs, target, FCBF_threshold = 0.1,
  output = "modules", verbose = FALSE)

find_cbf_modules(discretized_exprs, target, FCBF_threshold = 0.1,
  output = "modules", verbose = FALSE)
}
\arguments{
\item{FCBF_threshold}{A threshold for the minimum correlation (as determined by symettrical uncertainty)
between each variable and the class used for wrapped FCBF function. Defaults to 0.1.}

\item{output}{Form of the output. For an adjacency matrix, "matrix", for the list of modules, "modules".}

\item{verbose}{Adds verbosity. Defaults to FALSE.}

\item{fc}{A fcoex object containing a discretized expression table}

\item{fc}{A fcoex object containing a discretized expression table}

\item{FCBF_threshold}{A threshold for the minimum correlation (as determined by symettrical uncertainty)
between each variable and the class used for wrapped FCBF function. Defaults to 0.1.}

\item{verbose}{Adds verbosity. Defaults to FALSE.}
}
\value{
Returns a list with the CBF modules found or a adjacency matrix of the graph

Returns a list with the CBF modules found or a adjacency matrix of the graph
}
\description{
find_cbf_modules uses Symmetrical Uncertainty as a correlation measure and the FCBF algorithm to

find_cbf_modules uses Symmetrical Uncertainty as a correlation measure and the FCBF algorithm to
}
\details{
1 - Filter the gene list by correlations to a class (Step 1)

and

2 - Determine soft thresholds for coexpression to genes predominantly correlated to a class.

1 - Filter the gene list by correlations to a class (Step 1)

and

2 - Determine soft thresholds for coexpression to genes predominantly correlated to a class.
}
\examples{
 data(scDengue)
 exprs <- SummarizedExperiment::assay(scDengue, 'logcounts')
 discrete_expression <- as.data.frame(discretize_exprs(exprs))
 head(discrete_expression[,1:4])
 infection <- SummarizedExperiment::colData(scDengue)
 target <- infection$infection
 cbf_modules <- find_modules(discrete_expression,target, FCBF_threshold = 0.06)
}
