% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_nets}
\alias{get_nets}
\alias{get_nets,fcoex-method}
\title{Network visualization}
\usage{
get_nets(fc, n = 10, min_elements = 5, ...)

\S4method{get_nets}{fcoex}(fc, n = 10, min_elements = 5, ...)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{n}{number of nodes to label}

\item{min_elements}{Minimum number of elements in a module for it to be plotted. Defaults to 5.}

\item{...}{Optional parameters.}
}
\value{
Object of class \code{fcoex} with profile plots
}
\description{
Creates network visualizations based on the adjacency matrix
obtained with the find_cbf_modules method
}
\details{
This function was copied and adapted from the CEMiTool package.
The visualization of networks in this function is derivative of the 
intelectual work of CEMiTool's authors.
}
\examples{
library(SingleCellExperiment) 
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)
fc <- get_nets(fc)
}
