% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{data}
\name{gmt}
\alias{gmt}
\title{Module Overrepresentation Analysis}
\format{An object of class \code{data.frame} with 84466 rows and 2 columns.}
\usage{
gmt
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{gmt}{Object of class \code{data.frame} with 2 columns, one with
pathways and one with genes}

\item{verbose}{logical. Report analysis steps.}

\item{...}{Optional parameters.}
}
\value{
Object of class \code{fcoex}
}
\description{
Performs overrepresentation analysis for each co-expression module found.
}
\examples{
# Get example fcoex object
data(fc)
# Read gmt file
}
\seealso{
\code{\link{ora_data}}
}
\keyword{datasets}
