% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{idents}
\alias{idents}
\alias{idents,fcoex-method}
\title{Retrieves module identities from the recluster function}
\usage{
idents(fc)

\S4method{idents}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}
}
\value{
Named object of class \code{list} with clusterings derived
from the recluster function.
}
\description{
Retrieves module identities from the recluster function
}
\examples{
data("fc")
idents(fc)
}
\references{
Guangchuang Yu, Li-Gen Wang, Yanyan Han, Qing-Yu He. clusterProfiler:
an R package for comparing biological themes among gene clusters. OMICS:
A Journal of Integrative Biology. 2012, 16(5):284-287.
}
