% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mini_pbmc3k}
\alias{mini_pbmc3k}
\title{Processed subset of the pbmc3k dataset from PBMC genomics}
\format{An object of class \code{SingleCellExperiment}}
\source{
\href{https://bioconductor.org/packages/release/data/experiment/html/TENxPBMCData.html}
}
\usage{
data(mini_pbmc3k)
}
\description{
A subset with 600 sampled cells and the top 1700 variable 
genes from the  TENxPBMCData package pbmc3k dataset.
}
\details{
Preprocessed in accordance to OSCA (August 2019, https://osca.bioconductor.org/)

scater::normalized .
PCA with 50 components.
snn graph on the PCA space + louvain clustering to yield 8 clusters . 
UMAP already ran
}
\examples{
data(mini_pbmc3k)

}
\keyword{datasets}
