% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{mod_names}
\alias{mod_names}
\alias{mod_names,fcoex-method}
\title{Get module names in a fcoex object}
\usage{
mod_names(fc, include_NC = TRUE)

\S4method{mod_names}{fcoex}(fc, include_NC = TRUE)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}

\item{include_NC}{Logical. Whether or not to include "Not.Correlated"
module. Defaults to \code{TRUE}.}
}
\value{
Module names
}
\description{
Get module names in a fcoex object
}
\examples{
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)
mod_names(fc)
}
