% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{module_to_gmt}
\alias{module_to_gmt}
\title{Transform module genes list to a gmt file}
\usage{
module_to_gmt(fc, directory = "./Tables")
}
\arguments{
\item{fc}{A fcoex object.}
}
\value{
A .gmt file containing module genes in each row
}
\description{
Transform module genes list to a gmt file
}
\examples{
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
module_to_gmt(fc)
}
\keyword{internal}
