% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{new_fcoex}
\alias{new_fcoex}
\title{Create a fcoex object}
\usage{
new_fcoex(expr = data.frame(), target = vector())
}
\arguments{
\item{target}{Original target classes for the cells (\code{factor}).}

\item{expression}{Normalized gene expression table from single-cells \code{data.frame}.}
}
\value{
Object of class \code{fcoex}
}
\description{
Create a fcoex object
}
\examples{
# Create new fcoex object
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
}
