% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{ora_data}
\alias{ora_data}
\alias{ora_data,fcoex-method}
\title{Retrieve over representation analysis (ORA) results}
\usage{
ora_data(fc)

\S4method{ora_data}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}
}
\value{
Object of class \code{data.frame} with ORA data
}
\description{
Retrieve over representation analysis (ORA) results
}
\details{
This function returns the results of the \code{mod_ora} function
on the \code{fcoex} object. The ID column corresponds to pathways in
the gmt file for which  genes in the modules were enriched. The Count
column shows the number of genes in the module that are enriched for each
pathway. The GeneRatio column shows the proportion of  genes in the module
enriched for a given pathway out of all the genes in the module enriched
for any given pathway. The BgRatio column shows the proportion of genes
in a given pathway out of all the genes in the gmt file. For more details,
please refer to the \code{clusterProfiler} package documentation.

This function was ipsis litteris adapted from the CEMiTool package.
}
\examples{
data("fc")
ora_data(fc)
}
\references{
Guangchuang Yu, Li-Gen Wang, Yanyan Han, Qing-Yu He. clusterProfiler:
an R package for comparing biological themes among gene clusters. OMICS:
A Journal of Integrative Biology. 2012, 16(5):284-287.
}
