% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_interactions}
\alias{plot_interactions}
\alias{plot_interactions,fcoex-method}
\title{Network visualization}
\usage{
plot_interactions(fc, n = 10, min_elements = 5, ...)

\S4method{plot_interactions}{fcoex}(fc, n = 10, min_elements = 5, ...)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{n}{number of nodes to label}

\item{min_elements}{Minimum number of elements in a module for it to be plotted. Defaults to 5.}

\item{...}{Optional parameters.}
}
\value{
Object of class \code{fcoex} with profile plots
}
\description{
Creates a graph based on interactions provided
}
\examples{
library(SingleCellExperiment)  
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)
fc <- plot_interactions(fc)
}
