% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.plot_one_interaction}
\alias{.plot_one_interaction}
\title{Network visualization}
\usage{
.plot_one_interaction(adjacency_matrix, n, color, name)
}
\arguments{
\item{adjacency_matrix}{An adajcency matrix from the \code{fcoex} object.}

\item{n}{Number of genes to be shown}

\item{color}{Color of the module to be plotted}

\item{name}{Name of the module to be plotted}

\item{...}{Optional parameters.}
}
\value{
A ggplot2 ('gg') object
}
\description{
Creates a graph based on interactions provided
This function was copied and adapted from the CEMiTool package.
The visualization of networks in this function is derivative of the
intelectual work of CEMiTool's authors.
}
