% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ora}
\alias{plot_ora}
\alias{plot_ora,fcoex-method}
\title{ORA visualization}
\usage{
plot_ora(fc, n = 10, pv_cut = 0.05, ...)

\S4method{plot_ora}{fcoex}(fc, n = 10, pv_cut = 0.05, ...)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{n}{number of enrichments to show}

\item{pv_cut}{p-value significance cutoff. Default is 0.05.}

\item{...}{parameters to plot_ora_single}
}
\value{
Object of class \code{fcoex} with ORA plots
}
\description{
Creates a bar plot with the results of module overrepresentation analysis
}
\details{
This function was copied and adapted from the CEMiTool package.
The visualization in this function is derivative of the 
intelectual work of CEMiTool's authors.
}
\examples{
data("fc")
gmt_fname <- system.file("extdata", "pathways.gmt", package = "CEMiTool")
gmt_in <- pathwayPCA::read_gmt(gmt_fname)
fc <- mod_ora(fc, gmt_in)
fc <- plot_ora(fc)
}
