% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{recluster}
\alias{recluster}
\alias{recluster,fcoex-method}
\title{Recluster cells based on fcoex module composition}
\usage{
recluster(
  fc,
  hclust_method = "ward.D2",
  dist_method = "manhattan",
  k = 2,
  verbose = TRUE
)

\S4method{recluster}{fcoex}(
  fc,
  hclust_method = "ward.D2",
  dist_method = "manhattan",
  k = 2,
  verbose = TRUE
)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}

\item{hclust_method}{method for the hclust function. Defaults to "ward.D2".}

\item{dist_method}{method for the dist function. Defaults to "manhattan".}

\item{k}{desired number of clustes. Defaults to 2.}

\item{verbose}{Adds verbosity, defaults to TRUE.}
}
\value{
Object of class \code{data.frame} with new clusters
}
\description{
Recluster cells based on fcoex module composition
}
\examples{
data("fc")
fc <- recluster(fc)
}
