% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{show,fcoex-method}
\alias{show,fcoex-method}
\title{Print a fcoex object}
\usage{
\S4method{show}{fcoex}(object)
}
\arguments{
\item{object}{Object of class fcoex}
}
\value{
A fcoex object.
}
\description{
Print a fcoex object
}
