% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{show_plot}
\alias{show_plot}
\alias{show_plot,fcoex-method}
\title{Retrieve fcoex net plots}
\usage{
show_plot(fc)

\S4method{show_plot}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}
}
\value{
A plot corresponding to a fcoex analysis
#' @examples 
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)
show_net(fc)
}
\description{
Retrieve fcoex net plots
}
