% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{show_ora}
\alias{show_ora}
\alias{show_ora,fcoex-method}
\title{Retrieve fcoex ora plots}
\usage{
show_ora(fc)

\S4method{show_ora}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}
}
\value{
A plot corresponding to a fcoex analysis
}
\description{
Retrieve fcoex ora plots
}
\examples{
data("fc")
show_ora(fc)
}
