% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{show_plot}
\alias{show_plot}
\alias{show_plot,fcoex-method}
\title{Retrieve fcoex object plots}
\usage{
show_plot(fc, value)

\S4method{show_plot}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}
}
\value{
A plot corresponding to a fcoex analysis
}
\description{
Retrieve fcoex object plots
}
