% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{write_files}
\alias{write_files}
\alias{write_files,fcoex-method}
\title{Save the fcoex object in files}
\usage{
write_files(fc, ...)

\S4method{write_files}{fcoex}(fc, directory = "./Tables",
  force = FALSE)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}

\item{...}{Optional parameters}

\item{directory}{a directory}

\item{force}{if the directory exists the execution will not stop}
}
\value{
A directory containing fcoex results in files.
}
\description{
Save the fcoex object in files
}
\examples{
# Get example fcoex object
data(fc)
# Save fcoex results in files
write_files(fc, directory=".", force=TRUE)

}
