% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{find_cbf_modules}
\alias{find_cbf_modules}
\alias{find_cbf_modules,fcoex-method}
\title{data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)}
\usage{
find_cbf_modules(
  fc,
  n_genes_selected_in_first_step = NULL,
  FCBF_threshold = 0.1,
  verbose = TRUE,
  is_parallel = FALSE
)

\S4method{find_cbf_modules}{fcoex}(
  fc,
  n_genes_selected_in_first_step = NULL,
  FCBF_threshold = 0.1,
  verbose = TRUE,
  is_parallel = FALSE
)
}
\value{
Returns a list with the CBF modules found or a adjacency matrix of the graph
}
\description{
data("mini_pbmc3k")
targets <- colData(mini_pbmc3k)$clusters
exprs <- as.data.frame(assay(mini_pbmc3k, "logcounts"))
fc <- new_fcoex(exprs, targets)
fc <- discretize(fc)
fc <- find_cbf_modules(fc)
}
