% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{mod_colors}
\alias{mod_colors}
\alias{mod_colors,fcoex-method}
\title{Set module colors mod_colors attribute}
\usage{
mod_colors(fc)

\S4method{mod_colors}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}
}
\value{
A vector with color names.
}
\description{
Set module colors mod_colors attribute
}
\examples{
data("fc")
mod_colors(fc)
}
