% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{mod_gene_num}
\alias{mod_gene_num}
\alias{mod_gene_num,fcoex-method}
\title{Get the number of genes in modules in a fcoex object}
\usage{
mod_gene_num(fc, module = NULL)

\S4method{mod_gene_num}{fcoex}(fc, module = NULL)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}

\item{module}{Default is NULL. If a character string designating a 
module is#' given, the number of genes in that module is returned instead.}
}
\value{
The number of genes in module(s)
}
\description{
Get the number of genes in modules in a fcoex object
}
\examples{
data("fc")
mod_gene_num(fc, module = "TYROBP")
}
