% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{mod_ora}
\alias{mod_ora}
\alias{mod_ora,fcoex-method}
\title{# Run module overrepresentation analysis}
\usage{
mod_ora(fc, gmt, verbose = FALSE)

\S4method{mod_ora}{fcoex}(fc, gmt, verbose = FALSE)
}
\arguments{
\item{fc}{A fcoex object.}

\item{gmt}{A gmt file with gene sets for ora analysis}

\item{verbose}{Controls verbosity. Defaults to FALSE.}
}
\value{
A fcoex object containing over-representation analysis data
}
\description{
This function was modified from the CEMiTool package.
Chunks of code were retained "as is"
}
\examples{
data("fc")
gmt_fname <- system.file("extdata", "pathways.gmt", package = "CEMiTool")
gmt_in <- pathwayPCA::read_gmt(gmt_fname)
fc <- mod_ora(fc, gmt_in)
}
