% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{module_genes}
\alias{module_genes}
\alias{module_genes,fcoex-method}
\title{Get the module genes in a fcoex object}
\usage{
module_genes(fc, module = NULL)

\S4method{module_genes}{fcoex}(fc, module = NULL)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}

\item{module}{A character string with the name of the module of which
genes are to be returned. Defaults to \code{NULL}, which returns the full
list of genes and modules.}
}
\value{
Object of class \code{data.frame} containing genes and their
respective module
}
\description{
This function was copied and adapted from the CEMiTool package.
}
\examples{
data("fc")
module_genes(fc)
}
