% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\name{module_to_gmt}
\alias{module_to_gmt}
\title{Transform module genes list to a gmt file}
\usage{
module_to_gmt(fc, directory = "./Tables")
}
\arguments{
\item{fc}{A fcoex object.}
}
\value{
A .gmt file containing module genes in each row
}
\description{
Transform module genes list to a gmt file
}
\examples{
data("fc")
module_to_gmt(fc)
}
\keyword{internal}
