% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoex.R
\docType{methods}
\name{nmodules}
\alias{nmodules}
\alias{nmodules,fcoex-method}
\title{Get the number of modules in a fcoex object}
\usage{
nmodules(fc)

\S4method{nmodules}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}}
}
\value{
number of modules
}
\description{
Get the number of modules in a fcoex object
}
\examples{
# Get example fcoex object
data(fc)
# Get the number of modules
nmodules(fc)

}
