% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_interactions}
\alias{plot_interactions}
\alias{plot_interactions,fcoex-method}
\title{Network visualization}
\usage{
plot_interactions(fc, n = 10, min_elements = 5, ...)

\S4method{plot_interactions}{fcoex}(fc, n = 10, min_elements = 5, ...)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{n}{number of nodes to label}

\item{min_elements}{Minimum number of elements in a module for it to be plotted. Defaults to 5.}

\item{...}{Optional parameters.}
}
\value{
Object of class \code{fcoex} with profile plots
}
\description{
Creates a graph based on interactions provided
}
\examples{
# Get example fcoex object
data(fc)
# Get example gene interactions data
int <- system.file("extdata", "interactions.tsv", package = "fcoex")
int_df <- read.delim(int)
# Include interaction data into fcoex object
interactions_data(fc) <- int_df
# Plot resulting networks
fc <- plot_interactions(fc)
# Check resulting plot
show_plot(fc, "interaction")

}
