% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_ora}
\alias{plot_ora}
\alias{plot_ora,fcoex-method}
\title{ORA visualization}
\usage{
plot_ora(fc, ...)

\S4method{plot_ora}{fcoex}(fc, n = 10, pv_cut = 0.05, ...)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}

\item{...}{parameters to plot_ora_single}

\item{n}{number of modules to show}

\item{pv_cut}{p-value significance cutoff. Default is 0.05.}
}
\value{
Object of class \code{fcoex} with ORA plots
}
\description{
Creates a bar plot with the results of module overrepresentation analysis
}
\examples{
data("fc")
gmt_fname <- system.file("extdata", "pathways.gmt", package = "CEMiTool")
gmt_in <- read_gmt(gmt_fname)
fc <- mod_ora(fc, gmt_in)
fc <- plot_ora(fc)
}
