% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{show_net}
\alias{show_net}
\alias{show_net,fcoex-method}
\title{Retrieve fcoex net plots}
\usage{
show_net(fc)

\S4method{show_net}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}
}
\value{
A plot corresponding to a fcoex analysis
}
\description{
Retrieve fcoex net plots
}
\examples{
\dontrun{
data("fc")
show_net(fc)
}
}
