% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{show_ora}
\alias{show_ora}
\alias{show_ora,fcoex-method}
\title{Retrieve fcoex ora plots}
\usage{
show_ora(fc)

\S4method{show_ora}{fcoex}(fc)
}
\arguments{
\item{fc}{Object of class \code{fcoex}.}
}
\value{
A plot corresponding to a fcoex analysis
}
\description{
This function was copied and adapted from the CEMiTool package.
The visualization in this function is derivative of the 
intelectual work of CEMiTool's authors.
}
\examples{
data("fc")
show_ora(fc)
}
